requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Flask==3.1.2
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-secret-manager==2.24.0
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.74.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
requests==2.32.5
rsa==4.9.1
urllib3==2.5.0
Werkzeug==3.1.3
